Add initial aws iot API support
