+++ requirements.txt
@@ -2,3 +2,4 @@ aioconsole>=0.3.1
 aiohttp>=3.11.12
 websockets>=8.1
 async-timeout>=4.0.3
+paho-mqtt>=2.0.0
